rename main package to dieveg.demo
