+++ src/main/java/dieveg/demo/ProvaApplication.java
@@ -1,4 +1,4 @@
-package main;
+package dieveg.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
