+++ src/main/java/dieveg/demo/model/Competizione.java
@@ -1,12 +1,10 @@
-package main.model;
+package dieveg.demo.model;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
