+++ src/main/java/dieveg/demo/model/Utente.java
@@ -1,4 +1,4 @@
-package main.model;
+package dieveg.demo.model;
 
 import java.io.Serializable;
 import java.time.LocalDate;
@@ -8,7 +8,6 @@ import java.util.List;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
