+++ src/main/java/dieveg/demo/repository/CompetizioneRepository.java
@@ -1,10 +1,9 @@
-package main.repository;
+package dieveg.demo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import main.model.Competizione;
-import main.model.Utente;
+import dieveg.demo.model.Competizione;
 
 @Repository
 public interface CompetizioneRepository extends JpaRepository<Competizione, Long> {
