+++ src/main/java/dieveg/demo/repository/UtenteRepository.java
@@ -1,14 +1,12 @@
-package main.repository;
+package dieveg.demo.repository;
 
-import java.awt.geom.Area;
 import java.util.Optional;
 
-import org.springframework.boot.autoconfigure.security.SecurityProperties.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import main.model.Utente;
+import dieveg.demo.model.Utente;
 
 @Repository
 public interface UtenteRepository extends JpaRepository<Utente, Long> {
